+++ src/codegen/render/graphql/mod.rs
@@ -3,6 +3,7 @@ pub mod r#enum;
 pub mod enum_value_definition;
 pub mod field;
 pub mod gql_types;
+pub mod interfaces;
 pub mod object;
 pub mod scalars;
 pub mod union;